Split off metadata operations into MetadataFetcherTest
