+++ clients/src/test/java/org/apache/kafka/clients/consumer/internals/FetcherTest.java
@@ -17,9 +17,7 @@
 package org.apache.kafka.clients.consumer.internals;
 
 import org.apache.kafka.clients.ApiVersions;
-import org.apache.kafka.clients.ClientDnsLookup;
 import org.apache.kafka.clients.ClientRequest;
-import org.apache.kafka.clients.ClientUtils;
 import org.apache.kafka.clients.FetchSessionHandler;
 import org.apache.kafka.clients.Metadata;
 import org.apache.kafka.clients.MockClient;
@@ -27,9 +25,6 @@ import org.apache.kafka.clients.NetworkClient;
 import org.apache.kafka.clients.NodeApiVersions;
 import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.LogTruncationException;
-import org.apache.kafka.clients.consumer.OffsetAndMetadata;
-import org.apache.kafka.clients.consumer.OffsetAndTimestamp;
 import org.apache.kafka.clients.consumer.OffsetOutOfRangeException;
 import org.apache.kafka.clients.consumer.OffsetResetStrategy;
 import org.apache.kafka.common.Cluster;
@@ -38,27 +33,17 @@ import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.MetricName;
 import org.apache.kafka.common.MetricNameTemplate;
 import org.apache.kafka.common.Node;
-import org.apache.kafka.common.PartitionInfo;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.TopicIdPartition;
 import org.apache.kafka.common.Uuid;
-import org.apache.kafka.common.errors.InvalidTopicException;
 import org.apache.kafka.common.errors.RecordTooLargeException;
 import org.apache.kafka.common.errors.SerializationException;
-import org.apache.kafka.common.errors.TimeoutException;
 import org.apache.kafka.common.errors.TopicAuthorizationException;
 import org.apache.kafka.common.header.Header;
 import org.apache.kafka.common.header.internals.RecordHeader;
 import org.apache.kafka.common.internals.ClusterResourceListeners;
 import org.apache.kafka.common.message.FetchResponseData;
 import org.apache.kafka.common.message.ApiMessageType;
-import org.apache.kafka.common.message.ListOffsetsRequestData.ListOffsetsPartition;
-import org.apache.kafka.common.message.ListOffsetsRequestData.ListOffsetsTopic;
-import org.apache.kafka.common.message.ListOffsetsResponseData;
-import org.apache.kafka.common.message.ListOffsetsResponseData.ListOffsetsPartitionResponse;
-import org.apache.kafka.common.message.ListOffsetsResponseData.ListOffsetsTopicResponse;
-import org.apache.kafka.common.message.OffsetForLeaderEpochRequestData;
-import org.apache.kafka.common.message.OffsetForLeaderEpochRequestData.OffsetForLeaderPartition;
 import org.apache.kafka.common.message.OffsetForLeaderEpochResponseData;
 import org.apache.kafka.common.message.OffsetForLeaderEpochResponseData.EpochEndOffset;
 import org.apache.kafka.common.message.OffsetForLeaderEpochResponseData.OffsetForLeaderTopicResult;
@@ -84,15 +69,10 @@ import org.apache.kafka.common.record.RecordBatch;
 import org.apache.kafka.common.record.Records;
 import org.apache.kafka.common.record.SimpleRecord;
 import org.apache.kafka.common.record.TimestampType;
-import org.apache.kafka.common.requests.AbstractRequest;
 import org.apache.kafka.common.requests.ApiVersionsResponse;
 import org.apache.kafka.common.requests.FetchRequest;
 import org.apache.kafka.common.requests.FetchResponse;
-import org.apache.kafka.common.requests.ListOffsetsRequest;
-import org.apache.kafka.common.requests.ListOffsetsResponse;
-import org.apache.kafka.common.requests.MetadataRequest;
 import org.apache.kafka.common.requests.MetadataResponse;
-import org.apache.kafka.common.requests.OffsetsForLeaderEpochRequest;
 import org.apache.kafka.common.requests.OffsetsForLeaderEpochResponse;
 import org.apache.kafka.common.requests.RequestTestUtils;
 import org.apache.kafka.common.serialization.ByteArrayDeserializer;
@@ -137,7 +117,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
@@ -145,8 +124,6 @@ import static java.util.Collections.singleton;
 import static java.util.Collections.singletonList;
 import static java.util.Collections.singletonMap;
 import static org.apache.kafka.common.requests.FetchMetadata.INVALID_SESSION_ID;
-import static org.apache.kafka.common.requests.OffsetsForLeaderEpochResponse.UNDEFINED_EPOCH;
-import static org.apache.kafka.common.requests.OffsetsForLeaderEpochResponse.UNDEFINED_EPOCH_OFFSET;
 import static org.apache.kafka.common.utils.Utils.mkSet;
 import static org.apache.kafka.test.TestUtils.assertOptional;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
@@ -1449,8 +1426,6 @@ public class FetcherTest {
     public void testFetchOnCompletedFetchesForAllPausedPartitions() {
         buildFetcher();
 
-        Map<TopicPartition, List<ConsumerRecord<byte[], byte[]>>> fetchedRecords;
-
         assignFromUser(mkSet(tp0, tp1));
 
         // seek to tp0 and tp1 in two polls to generate 2 complete requests and responses
@@ -1914,669 +1889,6 @@ public class FetcherTest {
         assertEquals(0, subscriptions.position(tp0).offset);
     }
 
-    @Test
-    public void testUpdateFetchPositionNoOpWithPositionSet() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.seek(tp0, 5L);
-
-        metadataFetcher.resetOffsetsIfNeeded();
-        assertFalse(client.hasInFlightRequests());
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testUpdateFetchPositionResetToDefaultOffset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0);
-
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.EARLIEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testUpdateFetchPositionResetToLatestOffset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        client.updateMetadata(initialUpdateResponse);
-
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    /**
-     * Make sure the client behaves appropriately when receiving an exception for unavailable offsets
-     */
-    @Test
-    public void testFetchOffsetErrors() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Fail with OFFSET_NOT_AVAILABLE
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.OFFSET_NOT_AVAILABLE, 1L, 5L), false);
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0));
-
-        // Fail with LEADER_NOT_AVAILABLE
-        time.sleep(retryBackoffMs);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.LEADER_NOT_AVAILABLE, 1L, 5L), false);
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0));
-
-        // Back to normal
-        time.sleep(retryBackoffMs);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(Errors.NONE, 1L, 5L), false);
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertTrue(subscriptions.hasValidPosition(tp0));
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(subscriptions.position(tp0).offset, 5L);
-    }
-
-    @Test
-    public void testListOffsetSendsReadUncommitted() {
-        testListOffsetsSendsIsolationLevel(IsolationLevel.READ_UNCOMMITTED);
-    }
-
-    @Test
-    public void testListOffsetSendsReadCommitted() {
-        testListOffsetsSendsIsolationLevel(IsolationLevel.READ_COMMITTED);
-    }
-
-    private void testListOffsetsSendsIsolationLevel(IsolationLevel isolationLevel) {
-        buildFetcher(OffsetResetStrategy.EARLIEST, new ByteArrayDeserializer(), new ByteArrayDeserializer(),
-                Integer.MAX_VALUE, isolationLevel);
-
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        client.prepareResponse(body -> {
-            ListOffsetsRequest request = (ListOffsetsRequest) body;
-            return request.isolationLevel() == isolationLevel;
-        }, listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testResetOffsetsSkipsBlackedOutConnections() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.EARLIEST);
-
-        // Check that we skip sending the ListOffset request when the node is blacked out
-        client.updateMetadata(initialUpdateResponse);
-        Node node = initialUpdateResponse.brokers().iterator().next();
-        client.backoff(node, 500);
-        metadataFetcher.resetOffsetsIfNeeded();
-        assertEquals(0, consumerClient.pendingRequestCount());
-        consumerClient.pollNoWakeup();
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertEquals(OffsetResetStrategy.EARLIEST, subscriptions.resetStrategy(tp0));
-
-        time.sleep(500);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.EARLIEST_TIMESTAMP),
-                listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testUpdateFetchPositionResetToEarliestOffset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.EARLIEST);
-
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.EARLIEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testResetOffsetsMetadataRefresh() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // First fetch fails with stale metadata
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.NOT_LEADER_OR_FOLLOWER, 1L, 5L), false);
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-
-        // Expect a metadata refresh
-        client.prepareMetadataUpdate(initialUpdateResponse);
-        consumerClient.pollNoWakeup();
-        assertFalse(client.hasPendingMetadataUpdates());
-
-        // Next fetch succeeds
-        time.sleep(retryBackoffMs);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testListOffsetNoUpdateMissingEpoch() {
-        buildFetcher();
-
-        // Set up metadata with no leader epoch
-        subscriptions.assignFromUser(singleton(tp0));
-        MetadataResponse metadataWithNoLeaderEpochs = RequestTestUtils.metadataUpdateWithIds(
-                "kafka-cluster", 1, Collections.emptyMap(), singletonMap(topicName, 4), tp -> null, topicIds);
-        client.updateMetadata(metadataWithNoLeaderEpochs);
-
-        // Return a ListOffsets response with leaderEpoch=1, we should ignore it
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(tp0, Errors.NONE, 1L, 5L, 1));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        // Reset should be satisfied and no metadata update requested
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(metadata.updateRequested());
-        assertFalse(metadata.lastSeenLeaderEpoch(tp0).isPresent());
-    }
-
-    @Test
-    public void testListOffsetUpdateEpoch() {
-        buildFetcher();
-
-        // Set up metadata with leaderEpoch=1
-        subscriptions.assignFromUser(singleton(tp0));
-        MetadataResponse metadataWithLeaderEpochs = RequestTestUtils.metadataUpdateWithIds(
-                "kafka-cluster", 1, Collections.emptyMap(), singletonMap(topicName, 4), tp -> 1, topicIds);
-        client.updateMetadata(metadataWithLeaderEpochs);
-
-        // Reset offsets to trigger ListOffsets call
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Now we see a ListOffsets with leaderEpoch=2 epoch, we trigger a metadata update
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP, 1),
-                listOffsetResponse(tp0, Errors.NONE, 1L, 5L, 2));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(metadata.updateRequested());
-        assertOptional(metadata.lastSeenLeaderEpoch(tp0), epoch -> assertEquals((long) epoch, 2));
-    }
-
-    @Test
-    public void testUpdateFetchPositionDisconnect() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // First request gets a disconnect
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 5L), true);
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-
-        // Expect a metadata refresh
-        client.prepareMetadataUpdate(initialUpdateResponse);
-        consumerClient.pollNoWakeup();
-        assertFalse(client.hasPendingMetadataUpdates());
-
-        // No retry until the backoff passes
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(client.hasInFlightRequests());
-        assertFalse(subscriptions.hasValidPosition(tp0));
-
-        // Next one succeeds
-        time.sleep(retryBackoffMs);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testAssignmentChangeWithInFlightReset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Send the ListOffsets request to reset the position
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertTrue(client.hasInFlightRequests());
-
-        // Now we have an assignment change
-        assignFromUser(singleton(tp1));
-
-        // The response returns and is discarded
-        client.respond(listOffsetResponse(Errors.NONE, 1L, 5L));
-        consumerClient.pollNoWakeup();
-
-        assertFalse(client.hasPendingResponses());
-        assertFalse(client.hasInFlightRequests());
-        assertFalse(subscriptions.isAssigned(tp0));
-    }
-
-    @Test
-    public void testSeekWithInFlightReset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Send the ListOffsets request to reset the position
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertTrue(client.hasInFlightRequests());
-
-        // Now we get a seek from the user
-        subscriptions.seek(tp0, 237);
-
-        // The response returns and is discarded
-        client.respond(listOffsetResponse(Errors.NONE, 1L, 5L));
-        consumerClient.pollNoWakeup();
-
-        assertFalse(client.hasPendingResponses());
-        assertFalse(client.hasInFlightRequests());
-        assertEquals(237L, subscriptions.position(tp0).offset);
-    }
-
-    private boolean listOffsetMatchesExpectedReset(
-        TopicPartition tp,
-        OffsetResetStrategy strategy,
-        AbstractRequest request
-    ) {
-        assertTrue(request instanceof ListOffsetsRequest);
-
-        ListOffsetsRequest req = (ListOffsetsRequest) request;
-        assertEquals(singleton(tp.topic()), req.data().topics().stream()
-            .map(ListOffsetsTopic::name).collect(Collectors.toSet()));
-
-        ListOffsetsTopic listTopic = req.data().topics().get(0);
-        assertEquals(singleton(tp.partition()), listTopic.partitions().stream()
-            .map(ListOffsetsPartition::partitionIndex).collect(Collectors.toSet()));
-
-        ListOffsetsPartition listPartition = listTopic.partitions().get(0);
-        if (strategy == OffsetResetStrategy.EARLIEST) {
-            assertEquals(ListOffsetsRequest.EARLIEST_TIMESTAMP, listPartition.timestamp());
-        } else if (strategy == OffsetResetStrategy.LATEST) {
-            assertEquals(ListOffsetsRequest.LATEST_TIMESTAMP, listPartition.timestamp());
-        }
-        return true;
-    }
-
-    @Test
-    public void testEarlierOffsetResetArrivesLate() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.EARLIEST);
-        metadataFetcher.resetOffsetsIfNeeded();
-
-        client.prepareResponse(req -> {
-            if (listOffsetMatchesExpectedReset(tp0, OffsetResetStrategy.EARLIEST, req)) {
-                // Before the response is handled, we get a request to reset to the latest offset
-                subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-                return true;
-            } else {
-                return false;
-            }
-        }, listOffsetResponse(Errors.NONE, 1L, 0L));
-        consumerClient.pollNoWakeup();
-
-        // The list offset result should be ignored
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertEquals(OffsetResetStrategy.LATEST, subscriptions.resetStrategy(tp0));
-
-        metadataFetcher.resetOffsetsIfNeeded();
-        client.prepareResponse(
-            req -> listOffsetMatchesExpectedReset(tp0, OffsetResetStrategy.LATEST, req),
-            listOffsetResponse(Errors.NONE, 1L, 10L)
-        );
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertEquals(10, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testChangeResetWithInFlightReset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Send the ListOffsets request to reset the position
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertTrue(client.hasInFlightRequests());
-
-        // Now we get a seek from the user
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.EARLIEST);
-
-        // The response returns and is discarded
-        client.respond(listOffsetResponse(Errors.NONE, 1L, 5L));
-        consumerClient.pollNoWakeup();
-
-        assertFalse(client.hasPendingResponses());
-        assertFalse(client.hasInFlightRequests());
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertEquals(OffsetResetStrategy.EARLIEST, subscriptions.resetStrategy(tp0));
-    }
-
-    @Test
-    public void testIdempotentResetWithInFlightReset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Send the ListOffsets request to reset the position
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertTrue(client.hasInFlightRequests());
-
-        // Now we get a seek from the user
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        client.respond(listOffsetResponse(Errors.NONE, 1L, 5L));
-        consumerClient.pollNoWakeup();
-
-        assertFalse(client.hasInFlightRequests());
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertEquals(5L, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testRestOffsetsAuthorizationFailure() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // First request gets a disconnect
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.TOPIC_AUTHORIZATION_FAILED, -1, -1), false);
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.hasValidPosition(tp0));
-
-        try {
-            metadataFetcher.resetOffsetsIfNeeded();
-            fail("Expected authorization error to be raised");
-        } catch (TopicAuthorizationException e) {
-            assertEquals(singleton(tp0.topic()), e.unauthorizedTopics());
-        }
-
-        // The exception should clear after being raised, but no retry until the backoff
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-        assertFalse(client.hasInFlightRequests());
-        assertFalse(subscriptions.hasValidPosition(tp0));
-
-        // Next one succeeds
-        time.sleep(retryBackoffMs);
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(Errors.NONE, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertTrue(subscriptions.isFetchable(tp0));
-        assertEquals(5, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testFetchingPendingPartitionsBeforeAndAfterSubscriptionReset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.seek(tp0, 100);
-        assertEquals(100, subscriptions.position(tp0).offset);
-
-        assertTrue(subscriptions.isFetchable(tp0));
-
-        subscriptions.markPendingRevocation(singleton(tp0));
-        metadataFetcher.resetOffsetsIfNeeded();
-
-        // once a partition is marked pending, it should not be fetchable
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0));
-        assertTrue(subscriptions.hasValidPosition(tp0));
-        assertEquals(100, subscriptions.position(tp0).offset);
-
-        subscriptions.seek(tp0, 100);
-        assertEquals(100, subscriptions.position(tp0).offset);
-
-        // reassignment should enable fetching of the same partition
-        subscriptions.unsubscribe();
-        assignFromUser(singleton(tp0));
-        subscriptions.seek(tp0, 100);
-        assertEquals(100, subscriptions.position(tp0).offset);
-        assertTrue(subscriptions.isFetchable(tp0));
-    }
-
-    @Test
-    public void testUpdateFetchPositionOfPausedPartitionsRequiringOffsetReset() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.pause(tp0); // paused partition does not have a valid position
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP,
-            validLeaderEpoch), listOffsetResponse(Errors.NONE, 1L, 10L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0)); // because tp is paused
-        assertTrue(subscriptions.hasValidPosition(tp0));
-        assertEquals(10, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testUpdateFetchPositionOfPausedPartitionsWithoutAValidPosition() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0);
-        subscriptions.pause(tp0); // paused partition does not have a valid position
-
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0)); // because tp is paused
-        assertFalse(subscriptions.hasValidPosition(tp0));
-    }
-
-    @Test
-    public void testUpdateFetchPositionOfPausedPartitionsWithAValidPosition() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        subscriptions.seek(tp0, 10);
-        subscriptions.pause(tp0); // paused partition already has a valid position
-
-        metadataFetcher.resetOffsetsIfNeeded();
-
-        assertFalse(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0)); // because tp is paused
-        assertTrue(subscriptions.hasValidPosition(tp0));
-        assertEquals(10, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testGetAllTopics() {
-        // sending response before request, as getTopicMetadata is a blocking call
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(newMetadataResponse(topicName, Errors.NONE));
-
-        Map<String, List<PartitionInfo>> allTopics = metadataFetcher.getAllTopicMetadata(time.timer(5000L));
-
-        assertEquals(initialUpdateResponse.topicMetadata().size(), allTopics.size());
-    }
-
-    @Test
-    public void testGetAllTopicsDisconnect() {
-        // first try gets a disconnect, next succeeds
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(null, true);
-        client.prepareResponse(newMetadataResponse(topicName, Errors.NONE));
-        Map<String, List<PartitionInfo>> allTopics = metadataFetcher.getAllTopicMetadata(time.timer(5000L));
-        assertEquals(initialUpdateResponse.topicMetadata().size(), allTopics.size());
-    }
-
-    @Test
-    public void testGetAllTopicsTimeout() {
-        // since no response is prepared, the request should timeout
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        assertThrows(TimeoutException.class, () -> metadataFetcher.getAllTopicMetadata(time.timer(50L)));
-    }
-
-    @Test
-    public void testGetAllTopicsUnauthorized() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(newMetadataResponse(topicName, Errors.TOPIC_AUTHORIZATION_FAILED));
-        try {
-            metadataFetcher.getAllTopicMetadata(time.timer(10L));
-            fail();
-        } catch (TopicAuthorizationException e) {
-            assertEquals(singleton(topicName), e.unauthorizedTopics());
-        }
-    }
-
-    @Test
-    public void testGetTopicMetadataInvalidTopic() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(newMetadataResponse(topicName, Errors.INVALID_TOPIC_EXCEPTION));
-        assertThrows(InvalidTopicException.class, () -> metadataFetcher.getTopicMetadata(
-                new MetadataRequest.Builder(Collections.singletonList(topicName), true), time.timer(5000L)));
-    }
-
-    @Test
-    public void testGetTopicMetadataUnknownTopic() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(newMetadataResponse(topicName, Errors.UNKNOWN_TOPIC_OR_PARTITION));
-
-        Map<String, List<PartitionInfo>> topicMetadata = metadataFetcher.getTopicMetadata(
-                new MetadataRequest.Builder(Collections.singletonList(topicName), true), time.timer(5000L));
-        assertNull(topicMetadata.get(topicName));
-    }
-
-    @Test
-    public void testGetTopicMetadataLeaderNotAvailable() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(newMetadataResponse(topicName, Errors.LEADER_NOT_AVAILABLE));
-        client.prepareResponse(newMetadataResponse(topicName, Errors.NONE));
-
-        Map<String, List<PartitionInfo>> topicMetadata = metadataFetcher.getTopicMetadata(
-                new MetadataRequest.Builder(Collections.singletonList(topicName), true), time.timer(5000L));
-        assertTrue(topicMetadata.containsKey(topicName));
-    }
-
-    @Test
-    public void testGetTopicMetadataOfflinePartitions() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        MetadataResponse originalResponse = newMetadataResponse(topicName, Errors.NONE); //baseline ok response
-
-        //create a response based on the above one with all partitions being leaderless
-        List<MetadataResponse.TopicMetadata> altTopics = new ArrayList<>();
-        for (MetadataResponse.TopicMetadata item : originalResponse.topicMetadata()) {
-            List<MetadataResponse.PartitionMetadata> partitions = item.partitionMetadata();
-            List<MetadataResponse.PartitionMetadata> altPartitions = new ArrayList<>();
-            for (MetadataResponse.PartitionMetadata p : partitions) {
-                altPartitions.add(new MetadataResponse.PartitionMetadata(
-                    p.error,
-                    p.topicPartition,
-                    Optional.empty(), //no leader
-                    Optional.empty(),
-                    p.replicaIds,
-                    p.inSyncReplicaIds,
-                    p.offlineReplicaIds
-                ));
-            }
-            MetadataResponse.TopicMetadata alteredTopic = new MetadataResponse.TopicMetadata(
-                item.error(),
-                item.topic(),
-                item.isInternal(),
-                altPartitions
-            );
-            altTopics.add(alteredTopic);
-        }
-        Node controller = originalResponse.controller();
-        MetadataResponse altered = RequestTestUtils.metadataResponse(
-            originalResponse.brokers(),
-            originalResponse.clusterId(),
-            controller != null ? controller.id() : MetadataResponse.NO_CONTROLLER_ID,
-            altTopics);
-
-        client.prepareResponse(altered);
-
-        Map<String, List<PartitionInfo>> topicMetadata =
-            metadataFetcher.getTopicMetadata(new MetadataRequest.Builder(Collections.singletonList(topicName), false),
-                    time.timer(5000L));
-
-        assertNotNull(topicMetadata);
-        assertNotNull(topicMetadata.get(topicName));
-        //noinspection ConstantConditions
-        assertEquals(metadata.fetch().partitionCountForTopic(topicName).longValue(), topicMetadata.get(topicName).size());
-    }
-
     /*
      * Send multiple requests. Verify that the client side quota metrics have the right values
      */
@@ -2977,340 +2289,6 @@ public class FetcherTest {
         client.prepareResponse(fetchResponse(tp, records, error, hw, lastStableOffset, logStartOffset, throttleTime));
         consumerClient.poll(time.timer(0));
         return fetchedRecords();
-    }
-
-    @Test
-    public void testGetOffsetsForTimesTimeout() {
-        buildFetcher();
-        assertThrows(TimeoutException.class, () -> metadataFetcher.offsetsForTimes(
-            Collections.singletonMap(new TopicPartition(topicName, 2), 1000L), time.timer(100L)));
-    }
-
-    @Test
-    public void testGetOffsetsForTimes() {
-        buildFetcher();
-
-        // Empty map
-        assertTrue(metadataFetcher.offsetsForTimes(new HashMap<>(), time.timer(100L)).isEmpty());
-        // Unknown Offset
-        testGetOffsetsForTimesWithUnknownOffset();
-        // Error code none with unknown offset
-        testGetOffsetsForTimesWithError(Errors.NONE, Errors.NONE, -1L, 100L, null, 100L);
-        // Error code none with known offset
-        testGetOffsetsForTimesWithError(Errors.NONE, Errors.NONE, 10L, 100L, 10L, 100L);
-        // Test both of partition has error.
-        testGetOffsetsForTimesWithError(Errors.NOT_LEADER_OR_FOLLOWER, Errors.INVALID_REQUEST, 10L, 100L, 10L, 100L);
-        // Test the second partition has error.
-        testGetOffsetsForTimesWithError(Errors.NONE, Errors.NOT_LEADER_OR_FOLLOWER, 10L, 100L, 10L, 100L);
-        // Test different errors.
-        testGetOffsetsForTimesWithError(Errors.NOT_LEADER_OR_FOLLOWER, Errors.NONE, 10L, 100L, 10L, 100L);
-        testGetOffsetsForTimesWithError(Errors.UNKNOWN_TOPIC_OR_PARTITION, Errors.NONE, 10L, 100L, 10L, 100L);
-        testGetOffsetsForTimesWithError(Errors.UNSUPPORTED_FOR_MESSAGE_FORMAT, Errors.NONE, 10L, 100L, null, 100L);
-        testGetOffsetsForTimesWithError(Errors.BROKER_NOT_AVAILABLE, Errors.NONE, 10L, 100L, 10L, 100L);
-    }
-
-    @Test
-    public void testGetOffsetsFencedLeaderEpoch() {
-        buildFetcher();
-        subscriptions.assignFromUser(singleton(tp0));
-        client.updateMetadata(initialUpdateResponse);
-
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        client.prepareResponse(listOffsetResponse(Errors.FENCED_LEADER_EPOCH, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0));
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertEquals(0L, metadata.timeToNextUpdate(time.milliseconds()));
-    }
-
-    @Test
-    public void testGetOffsetByTimeWithPartitionsRetryCouldTriggerMetadataUpdate() {
-        List<Errors> retriableErrors = Arrays.asList(Errors.NOT_LEADER_OR_FOLLOWER,
-            Errors.REPLICA_NOT_AVAILABLE, Errors.KAFKA_STORAGE_ERROR, Errors.OFFSET_NOT_AVAILABLE,
-            Errors.LEADER_NOT_AVAILABLE, Errors.FENCED_LEADER_EPOCH, Errors.UNKNOWN_LEADER_EPOCH);
-
-        final int newLeaderEpoch = 3;
-        MetadataResponse updatedMetadata = RequestTestUtils.metadataUpdateWithIds("dummy", 3,
-            singletonMap(topicName, Errors.NONE), singletonMap(topicName, 4), tp -> newLeaderEpoch, topicIds);
-
-        Node originalLeader = initialUpdateResponse.buildCluster().leaderFor(tp1);
-        Node newLeader = updatedMetadata.buildCluster().leaderFor(tp1);
-        assertNotEquals(originalLeader, newLeader);
-
-        for (Errors retriableError : retriableErrors) {
-            buildFetcher();
-
-            subscriptions.assignFromUser(mkSet(tp0, tp1));
-            client.updateMetadata(initialUpdateResponse);
-
-            final long fetchTimestamp = 10L;
-            ListOffsetsPartitionResponse tp0NoError = new ListOffsetsPartitionResponse()
-                .setPartitionIndex(tp0.partition())
-                .setErrorCode(Errors.NONE.code())
-                .setTimestamp(fetchTimestamp)
-                .setOffset(4L);
-            List<ListOffsetsTopicResponse> topics = Collections.singletonList(
-                    new ListOffsetsTopicResponse()
-                        .setName(tp0.topic())
-                        .setPartitions(Arrays.asList(
-                                tp0NoError,
-                                new ListOffsetsPartitionResponse()
-                                    .setPartitionIndex(tp1.partition())
-                                    .setErrorCode(retriableError.code())
-                                    .setTimestamp(ListOffsetsRequest.LATEST_TIMESTAMP)
-                                    .setOffset(-1L))));
-            ListOffsetsResponseData data = new ListOffsetsResponseData()
-                    .setThrottleTimeMs(0)
-                    .setTopics(topics);
-
-            client.prepareResponseFrom(body -> {
-                boolean isListOffsetRequest = body instanceof ListOffsetsRequest;
-                if (isListOffsetRequest) {
-                    ListOffsetsRequest request = (ListOffsetsRequest) body;
-                    List<ListOffsetsTopic> expectedTopics = Collections.singletonList(
-                            new ListOffsetsTopic()
-                                .setName(tp0.topic())
-                                .setPartitions(Arrays.asList(
-                                        new ListOffsetsPartition()
-                                            .setPartitionIndex(tp1.partition())
-                                            .setTimestamp(fetchTimestamp)
-                                            .setCurrentLeaderEpoch(ListOffsetsResponse.UNKNOWN_EPOCH),
-                                        new ListOffsetsPartition()
-                                            .setPartitionIndex(tp0.partition())
-                                            .setTimestamp(fetchTimestamp)
-                                            .setCurrentLeaderEpoch(ListOffsetsResponse.UNKNOWN_EPOCH))));
-                    return request.topics().equals(expectedTopics);
-                } else {
-                    return false;
-                }
-            }, new ListOffsetsResponse(data), originalLeader);
-
-            client.prepareMetadataUpdate(updatedMetadata);
-
-            // If the metadata wasn't updated before retrying, the fetcher would consult the original leader and hit a NOT_LEADER exception.
-            // We will count the answered future response in the end to verify if this is the case.
-            List<ListOffsetsTopicResponse> topicsWithFatalError = Collections.singletonList(
-                    new ListOffsetsTopicResponse()
-                        .setName(tp0.topic())
-                        .setPartitions(Arrays.asList(
-                                tp0NoError,
-                                new ListOffsetsPartitionResponse()
-                                    .setPartitionIndex(tp1.partition())
-                                    .setErrorCode(Errors.NOT_LEADER_OR_FOLLOWER.code())
-                                    .setTimestamp(ListOffsetsRequest.LATEST_TIMESTAMP)
-                                    .setOffset(-1L))));
-            ListOffsetsResponseData dataWithFatalError = new ListOffsetsResponseData()
-                    .setThrottleTimeMs(0)
-                    .setTopics(topicsWithFatalError);
-            client.prepareResponseFrom(new ListOffsetsResponse(dataWithFatalError), originalLeader);
-
-            // The request to new leader must only contain one partition tp1 with error.
-            client.prepareResponseFrom(body -> {
-                boolean isListOffsetRequest = body instanceof ListOffsetsRequest;
-                if (isListOffsetRequest) {
-                    ListOffsetsRequest request = (ListOffsetsRequest) body;
-
-                    ListOffsetsTopic requestTopic = request.topics().get(0);
-                    ListOffsetsPartition expectedPartition = new ListOffsetsPartition()
-                            .setPartitionIndex(tp1.partition())
-                            .setTimestamp(fetchTimestamp)
-                            .setCurrentLeaderEpoch(newLeaderEpoch);
-                    return expectedPartition.equals(requestTopic.partitions().get(0));
-                } else {
-                    return false;
-                }
-            }, listOffsetResponse(tp1, Errors.NONE, fetchTimestamp, 5L), newLeader);
-
-            Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                metadataFetcher.offsetsForTimes(
-                    Utils.mkMap(Utils.mkEntry(tp0, fetchTimestamp),
-                    Utils.mkEntry(tp1, fetchTimestamp)), time.timer(Integer.MAX_VALUE));
-
-            assertEquals(Utils.mkMap(
-                Utils.mkEntry(tp0, new OffsetAndTimestamp(4L, fetchTimestamp)),
-                Utils.mkEntry(tp1, new OffsetAndTimestamp(5L, fetchTimestamp))), offsetAndTimestampMap);
-
-            // The NOT_LEADER exception future should not be cleared as we already refreshed the metadata before
-            // first retry, thus never hitting.
-            assertEquals(1, client.numAwaitingResponses());
-
-            fetcher.close();
-        }
-    }
-
-    @Test
-    public void testGetOffsetsUnknownLeaderEpoch() {
-        buildFetcher();
-        subscriptions.assignFromUser(singleton(tp0));
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        client.prepareResponse(listOffsetResponse(Errors.UNKNOWN_LEADER_EPOCH, 1L, 5L));
-        metadataFetcher.resetOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertTrue(subscriptions.isOffsetResetNeeded(tp0));
-        assertFalse(subscriptions.isFetchable(tp0));
-        assertFalse(subscriptions.hasValidPosition(tp0));
-        assertEquals(0L, metadata.timeToNextUpdate(time.milliseconds()));
-    }
-
-    @Test
-    public void testGetOffsetsIncludesLeaderEpoch() {
-        buildFetcher();
-        subscriptions.assignFromUser(singleton(tp0));
-
-        client.updateMetadata(initialUpdateResponse);
-
-        // Metadata update with leader epochs
-        MetadataResponse metadataResponse = RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), Collections.singletonMap(topicName, 4), tp -> 99, topicIds);
-        client.updateMetadata(metadataResponse);
-
-        // Request latest offset
-        subscriptions.requestOffsetReset(tp0);
-        metadataFetcher.resetOffsetsIfNeeded();
-
-        // Check for epoch in outgoing request
-        MockClient.RequestMatcher matcher = body -> {
-            if (body instanceof ListOffsetsRequest) {
-                ListOffsetsRequest offsetRequest = (ListOffsetsRequest) body;
-                int epoch = offsetRequest.topics().get(0).partitions().get(0).currentLeaderEpoch();
-                assertTrue(epoch != ListOffsetsResponse.UNKNOWN_EPOCH, "Expected Fetcher to set leader epoch in request");
-                assertEquals(epoch, 99, "Expected leader epoch to match epoch from metadata update");
-                return true;
-            } else {
-                fail("Should have seen ListOffsetRequest");
-                return false;
-            }
-        };
-
-        client.prepareResponse(matcher, listOffsetResponse(Errors.NONE, 1L, 5L));
-        consumerClient.pollNoWakeup();
-    }
-
-    @Test
-    public void testGetOffsetsForTimesWhenSomeTopicPartitionLeadersNotKnownInitially() {
-        buildFetcher();
-
-        subscriptions.assignFromUser(mkSet(tp0, tp1));
-        final String anotherTopic = "another-topic";
-        final TopicPartition t2p0 = new TopicPartition(anotherTopic, 0);
-
-        client.reset();
-
-        // Metadata initially has one topic
-        MetadataResponse initialMetadata = RequestTestUtils.metadataUpdateWithIds(3, singletonMap(topicName, 2), topicIds);
-        client.updateMetadata(initialMetadata);
-
-        // The first metadata refresh should contain one topic
-        client.prepareMetadataUpdate(initialMetadata);
-        client.prepareResponseFrom(listOffsetResponse(tp0, Errors.NONE, 1000L, 11L),
-                metadata.fetch().leaderFor(tp0));
-        client.prepareResponseFrom(listOffsetResponse(tp1, Errors.NONE, 1000L, 32L),
-                metadata.fetch().leaderFor(tp1));
-
-        // Second metadata refresh should contain two topics
-        Map<String, Integer> partitionNumByTopic = new HashMap<>();
-        partitionNumByTopic.put(topicName, 2);
-        partitionNumByTopic.put(anotherTopic, 1);
-        topicIds.put("another-topic", Uuid.randomUuid());
-        MetadataResponse updatedMetadata = RequestTestUtils.metadataUpdateWithIds(3, partitionNumByTopic, topicIds);
-        client.prepareMetadataUpdate(updatedMetadata);
-        client.prepareResponseFrom(listOffsetResponse(t2p0, Errors.NONE, 1000L, 54L),
-                metadata.fetch().leaderFor(t2p0));
-
-        Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
-        timestampToSearch.put(tp0, ListOffsetsRequest.LATEST_TIMESTAMP);
-        timestampToSearch.put(tp1, ListOffsetsRequest.LATEST_TIMESTAMP);
-        timestampToSearch.put(t2p0, ListOffsetsRequest.LATEST_TIMESTAMP);
-        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-            metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
-
-        assertNotNull(offsetAndTimestampMap.get(tp0), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + tp0);
-        assertNotNull(offsetAndTimestampMap.get(tp1), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + tp1);
-        assertNotNull(offsetAndTimestampMap.get(t2p0), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + t2p0);
-        assertEquals(11L, offsetAndTimestampMap.get(tp0).offset());
-        assertEquals(32L, offsetAndTimestampMap.get(tp1).offset());
-        assertEquals(54L, offsetAndTimestampMap.get(t2p0).offset());
-    }
-
-    @Test
-    public void testGetOffsetsForTimesWhenSomeTopicPartitionLeadersDisconnectException() {
-        buildFetcher();
-        final String anotherTopic = "another-topic";
-        final TopicPartition t2p0 = new TopicPartition(anotherTopic, 0);
-        subscriptions.assignFromUser(mkSet(tp0, t2p0));
-
-        client.reset();
-
-        MetadataResponse initialMetadata = RequestTestUtils.metadataUpdateWithIds(1, singletonMap(topicName, 1), topicIds);
-        client.updateMetadata(initialMetadata);
-
-        Map<String, Integer> partitionNumByTopic = new HashMap<>();
-        partitionNumByTopic.put(topicName, 1);
-        partitionNumByTopic.put(anotherTopic, 1);
-        topicIds.put("another-topic", Uuid.randomUuid());
-        MetadataResponse updatedMetadata = RequestTestUtils.metadataUpdateWithIds(1, partitionNumByTopic, topicIds);
-        client.prepareMetadataUpdate(updatedMetadata);
-
-        client.prepareResponse(listOffsetRequestMatcher(ListOffsetsRequest.LATEST_TIMESTAMP),
-                listOffsetResponse(tp0, Errors.NONE, 1000L, 11L), true);
-        client.prepareResponseFrom(listOffsetResponse(tp0, Errors.NONE, 1000L, 11L), metadata.fetch().leaderFor(tp0));
-
-        Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
-        timestampToSearch.put(tp0, ListOffsetsRequest.LATEST_TIMESTAMP);
-        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap = metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
-
-        assertNotNull(offsetAndTimestampMap.get(tp0), "Expect MetadataFetcher.offsetsForTimes() to return non-null result for " + tp0);
-        assertEquals(11L, offsetAndTimestampMap.get(tp0).offset());
-        assertNotNull(metadata.fetch().partitionCountForTopic(anotherTopic));
-    }
-
-    @Test
-    public void testListOffsetsWithZeroTimeout() {
-        buildFetcher();
-
-        Map<TopicPartition, Long> offsetsToSearch = new HashMap<>();
-        offsetsToSearch.put(tp0, ListOffsetsRequest.EARLIEST_TIMESTAMP);
-        offsetsToSearch.put(tp1, ListOffsetsRequest.EARLIEST_TIMESTAMP);
-
-        Map<TopicPartition, Long> offsetsToExpect = new HashMap<>();
-        offsetsToExpect.put(tp0, null);
-        offsetsToExpect.put(tp1, null);
-
-        assertEquals(offsetsToExpect, metadataFetcher.offsetsForTimes(offsetsToSearch, time.timer(0)));
-    }
-
-    @Test
-    public void testBatchedListOffsetsMetadataErrors() {
-        buildFetcher();
-
-        ListOffsetsResponseData data = new ListOffsetsResponseData()
-                .setThrottleTimeMs(0)
-                .setTopics(Collections.singletonList(new ListOffsetsTopicResponse()
-                        .setName(tp0.topic())
-                        .setPartitions(Arrays.asList(
-                                new ListOffsetsPartitionResponse()
-                                    .setPartitionIndex(tp0.partition())
-                                    .setErrorCode(Errors.NOT_LEADER_OR_FOLLOWER.code())
-                                    .setTimestamp(ListOffsetsResponse.UNKNOWN_TIMESTAMP)
-                                    .setOffset(ListOffsetsResponse.UNKNOWN_OFFSET),
-                                new ListOffsetsPartitionResponse()
-                                    .setPartitionIndex(tp1.partition())
-                                    .setErrorCode(Errors.UNKNOWN_TOPIC_OR_PARTITION.code())
-                                    .setTimestamp(ListOffsetsResponse.UNKNOWN_TIMESTAMP)
-                                    .setOffset(ListOffsetsResponse.UNKNOWN_OFFSET)))));
-        client.prepareResponse(new ListOffsetsResponse(data));
-
-        Map<TopicPartition, Long> offsetsToSearch = new HashMap<>();
-        offsetsToSearch.put(tp0, ListOffsetsRequest.EARLIEST_TIMESTAMP);
-        offsetsToSearch.put(tp1, ListOffsetsRequest.EARLIEST_TIMESTAMP);
-
-        assertThrows(TimeoutException.class, () -> metadataFetcher.offsetsForTimes(offsetsToSearch, time.timer(1)));
     }
 
     @Test
@@ -3913,15 +2891,6 @@ public class FetcherTest {
                 }
             }
         };
-        this.metadataFetcher = new MetadataFetcher(logContext,
-                                                   consumerClient,
-                                                   metadata,
-                                                   subscriptions,
-                                                   time,
-                                                   retryBackoffMs,
-                                                   requestTimeoutMs,
-                                                   isolationLevel,
-                                                   apiVersions);
 
         MetadataResponse initialMetadataResponse = RequestTestUtils.metadataUpdateWithIds(1,
                 singletonMap(topicName, numPartitions), tp -> validLeaderEpoch, topicIds);
@@ -4123,125 +3092,6 @@ public class FetcherTest {
         return 1;
     }
 
-    private void testGetOffsetsForTimesWithError(Errors errorForP0,
-                                                 Errors errorForP1,
-                                                 long offsetForP0,
-                                                 long offsetForP1,
-                                                 Long expectedOffsetForP0,
-                                                 Long expectedOffsetForP1) {
-        client.reset();
-        String topicName2 = "topic2";
-        TopicPartition t2p0 = new TopicPartition(topicName2, 0);
-        // Expect a metadata refresh.
-        metadata.bootstrap(ClientUtils.parseAndValidateAddresses(Collections.singletonList("1.1.1.1:1111"),
-                ClientDnsLookup.USE_ALL_DNS_IPS));
-
-        Map<String, Integer> partitionNumByTopic = new HashMap<>();
-        partitionNumByTopic.put(topicName, 2);
-        partitionNumByTopic.put(topicName2, 1);
-        MetadataResponse updateMetadataResponse = RequestTestUtils.metadataUpdateWithIds(2, partitionNumByTopic, topicIds);
-        Cluster updatedCluster = updateMetadataResponse.buildCluster();
-
-        // The metadata refresh should contain all the topics.
-        client.prepareMetadataUpdate(updateMetadataResponse, true);
-
-        // First try should fail due to metadata error.
-        client.prepareResponseFrom(listOffsetResponse(t2p0, errorForP0, offsetForP0, offsetForP0),
-                updatedCluster.leaderFor(t2p0));
-        client.prepareResponseFrom(listOffsetResponse(tp1, errorForP1, offsetForP1, offsetForP1),
-                updatedCluster.leaderFor(tp1));
-        // Second try should succeed.
-        client.prepareResponseFrom(listOffsetResponse(t2p0, Errors.NONE, offsetForP0, offsetForP0),
-                updatedCluster.leaderFor(t2p0));
-        client.prepareResponseFrom(listOffsetResponse(tp1, Errors.NONE, offsetForP1, offsetForP1),
-                updatedCluster.leaderFor(tp1));
-
-        Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
-        timestampToSearch.put(t2p0, 0L);
-        timestampToSearch.put(tp1, 0L);
-        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
-
-        if (expectedOffsetForP0 == null)
-            assertNull(offsetAndTimestampMap.get(t2p0));
-        else {
-            assertEquals(expectedOffsetForP0.longValue(), offsetAndTimestampMap.get(t2p0).timestamp());
-            assertEquals(expectedOffsetForP0.longValue(), offsetAndTimestampMap.get(t2p0).offset());
-        }
-
-        if (expectedOffsetForP1 == null)
-            assertNull(offsetAndTimestampMap.get(tp1));
-        else {
-            assertEquals(expectedOffsetForP1.longValue(), offsetAndTimestampMap.get(tp1).timestamp());
-            assertEquals(expectedOffsetForP1.longValue(), offsetAndTimestampMap.get(tp1).offset());
-        }
-    }
-
-    private void testGetOffsetsForTimesWithUnknownOffset() {
-        client.reset();
-        // Ensure metadata has both partitions.
-        MetadataResponse initialMetadataUpdate = RequestTestUtils.metadataUpdateWithIds(1, singletonMap(topicName, 1), topicIds);
-        client.updateMetadata(initialMetadataUpdate);
-
-        ListOffsetsResponseData data = new ListOffsetsResponseData()
-                .setThrottleTimeMs(0)
-                .setTopics(Collections.singletonList(new ListOffsetsTopicResponse()
-                        .setName(tp0.topic())
-                        .setPartitions(Collections.singletonList(new ListOffsetsPartitionResponse()
-                                .setPartitionIndex(tp0.partition())
-                                .setErrorCode(Errors.NONE.code())
-                                .setTimestamp(ListOffsetsResponse.UNKNOWN_TIMESTAMP)
-                                .setOffset(ListOffsetsResponse.UNKNOWN_OFFSET)))));
-
-        client.prepareResponseFrom(new ListOffsetsResponse(data),
-                metadata.fetch().leaderFor(tp0));
-
-        Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
-        timestampToSearch.put(tp0, 0L);
-        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
-
-        assertTrue(offsetAndTimestampMap.containsKey(tp0));
-        assertNull(offsetAndTimestampMap.get(tp0));
-    }
-
-    @Test
-    public void testGetOffsetsForTimesWithUnknownOffsetV0() {
-        buildFetcher();
-        // Empty map
-        assertTrue(metadataFetcher.offsetsForTimes(new HashMap<>(), time.timer(100L)).isEmpty());
-        // Unknown Offset
-        client.reset();
-        // Ensure metadata has both partition.
-        MetadataResponse initialMetadataUpdate = RequestTestUtils.metadataUpdateWithIds(1, singletonMap(topicName, 1), topicIds);
-        client.updateMetadata(initialMetadataUpdate);
-        // Force LIST_OFFSETS version 0
-        Node node = metadata.fetch().nodes().get(0);
-        apiVersions.update(node.idString(), NodeApiVersions.create(
-            ApiKeys.LIST_OFFSETS.id, (short) 0, (short) 0));
-
-        ListOffsetsResponseData data = new ListOffsetsResponseData()
-                .setThrottleTimeMs(0)
-                .setTopics(Collections.singletonList(new ListOffsetsTopicResponse()
-                        .setName(tp0.topic())
-                        .setPartitions(Collections.singletonList(new ListOffsetsPartitionResponse()
-                                .setPartitionIndex(tp0.partition())
-                                .setErrorCode(Errors.NONE.code())
-                                .setTimestamp(ListOffsetsResponse.UNKNOWN_TIMESTAMP)
-                                .setOldStyleOffsets(Collections.emptyList())))));
-
-        client.prepareResponseFrom(new ListOffsetsResponse(data),
-                metadata.fetch().leaderFor(tp0));
-
-        Map<TopicPartition, Long> timestampToSearch = new HashMap<>();
-        timestampToSearch.put(tp0, 0L);
-        Map<TopicPartition, OffsetAndTimestamp> offsetAndTimestampMap =
-                metadataFetcher.offsetsForTimes(timestampToSearch, time.timer(Long.MAX_VALUE));
-
-        assertTrue(offsetAndTimestampMap.containsKey(tp0));
-        assertNull(offsetAndTimestampMap.get(tp0));
-    }
-
     @Test
     public void testSubscriptionPositionUpdatedWithEpoch() {
         // Create some records that include a leader epoch (1)
@@ -4291,395 +3141,6 @@ public class FetcherTest {
     }
 
     @Test
-    public void testOffsetValidationRequestGrouping() {
-        buildFetcher();
-        assignFromUser(mkSet(tp0, tp1, tp2, tp3));
-
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 3,
-            Collections.emptyMap(), singletonMap(topicName, 4),
-            tp -> 5, topicIds), false, 0L);
-
-        for (TopicPartition tp : subscriptions.assignedPartitions()) {
-            Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(
-                metadata.currentLeader(tp).leader, Optional.of(4));
-            subscriptions.seekUnvalidated(tp,
-                new SubscriptionState.FetchPosition(0, Optional.of(4), leaderAndEpoch));
-        }
-
-        Set<TopicPartition> allRequestedPartitions = new HashSet<>();
-
-        for (Node node : metadata.fetch().nodes()) {
-            apiVersions.update(node.idString(), NodeApiVersions.create());
-
-            Set<TopicPartition> expectedPartitions = subscriptions.assignedPartitions().stream()
-                .filter(tp ->
-                    metadata.currentLeader(tp).leader.equals(Optional.of(node)))
-                .collect(Collectors.toSet());
-
-            assertTrue(expectedPartitions.stream().noneMatch(allRequestedPartitions::contains));
-            assertTrue(expectedPartitions.size() > 0);
-            allRequestedPartitions.addAll(expectedPartitions);
-
-            OffsetForLeaderEpochResponseData data = new OffsetForLeaderEpochResponseData();
-            expectedPartitions.forEach(tp -> {
-                OffsetForLeaderTopicResult topic = data.topics().find(tp.topic());
-                if (topic == null) {
-                    topic = new OffsetForLeaderTopicResult().setTopic(tp.topic());
-                    data.topics().add(topic);
-                }
-                topic.partitions().add(new EpochEndOffset()
-                    .setPartition(tp.partition())
-                    .setErrorCode(Errors.NONE.code())
-                    .setLeaderEpoch(4)
-                    .setEndOffset(0));
-            });
-
-            OffsetsForLeaderEpochResponse response = new OffsetsForLeaderEpochResponse(data);
-            client.prepareResponseFrom(body -> {
-                OffsetsForLeaderEpochRequest request = (OffsetsForLeaderEpochRequest) body;
-                return expectedPartitions.equals(offsetForLeaderPartitionMap(request.data()).keySet());
-            }, response, node);
-        }
-
-        assertEquals(subscriptions.assignedPartitions(), allRequestedPartitions);
-
-        metadataFetcher.validateOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertTrue(subscriptions.assignedPartitions()
-            .stream().noneMatch(subscriptions::awaitingValidation));
-    }
-
-    @Test
-    public void testOffsetValidationAwaitsNodeApiVersion() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        final int epochOne = 1;
-
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), partitionCounts, tp -> epochOne, topicIds), false, 0L);
-
-        Node node = metadata.fetch().nodes().get(0);
-        assertFalse(client.isConnected(node.idString()));
-
-        // Seek with a position and leader+epoch
-        Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(
-                metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-        subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(20L, Optional.of(epochOne), leaderAndEpoch));
-        assertFalse(client.isConnected(node.idString()));
-        assertTrue(subscriptions.awaitingValidation(tp0));
-
-        // No version information is initially available, but the node is now connected
-        metadataFetcher.validateOffsetsIfNeeded();
-        assertTrue(subscriptions.awaitingValidation(tp0));
-        assertTrue(client.isConnected(node.idString()));
-        apiVersions.update(node.idString(), NodeApiVersions.create());
-
-        // On the next call, the OffsetForLeaderEpoch request is sent and validation completes
-        client.prepareResponseFrom(
-            prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, epochOne, 30L),
-            node);
-
-        metadataFetcher.validateOffsetsIfNeeded();
-        consumerClient.pollNoWakeup();
-
-        assertFalse(subscriptions.awaitingValidation(tp0));
-        assertEquals(20L, subscriptions.position(tp0).offset);
-    }
-
-    @Test
-    public void testOffsetValidationSkippedForOldBroker() {
-        // Old brokers may require CLUSTER permission to use the OffsetForLeaderEpoch API,
-        // so we should skip offset validation and not send the request.
-
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        final int epochOne = 1;
-        final int epochTwo = 2;
-
-        // Start with metadata, epoch=1
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), partitionCounts, tp -> epochOne, topicIds), false, 0L);
-
-        // Offset validation requires OffsetForLeaderEpoch request v3 or higher
-        Node node = metadata.fetch().nodes().get(0);
-        apiVersions.update(node.idString(), NodeApiVersions.create(
-            ApiKeys.OFFSET_FOR_LEADER_EPOCH.id, (short) 0, (short) 2));
-
-        {
-            // Seek with a position and leader+epoch
-            Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(
-                    metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-            subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(epochOne), leaderAndEpoch));
-
-            // Update metadata to epoch=2, enter validation
-            metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                    Collections.emptyMap(), partitionCounts, tp -> epochTwo, topicIds), false, 0L);
-            metadataFetcher.validateOffsetsIfNeeded();
-
-            // Offset validation is skipped
-            assertFalse(subscriptions.awaitingValidation(tp0));
-        }
-
-        {
-            // Seek with a position and leader+epoch
-            Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(
-                    metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-            subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(epochOne), leaderAndEpoch));
-
-            // Update metadata to epoch=2, enter validation
-            metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                    Collections.emptyMap(), partitionCounts, tp -> epochTwo, topicIds), false, 0L);
-
-            // Subscription should not stay in AWAITING_VALIDATION in prepareFetchRequest
-            assertEquals(1, fetcher.sendFetches());
-            assertFalse(subscriptions.awaitingValidation(tp0));
-        }
-    }
-
-    @Test
-    public void testOffsetValidationSkippedForOldResponse() {
-        // Old responses may provide unreliable leader epoch,
-        // so we should skip offset validation and not send the request.
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        final int epochOne = 1;
-
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-            Collections.emptyMap(), partitionCounts, tp -> epochOne, topicIds), false, 0L);
-
-        Node node = metadata.fetch().nodes().get(0);
-        assertFalse(client.isConnected(node.idString()));
-
-        // Seek with a position and leader+epoch
-        Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(
-            metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-        subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(20L, Optional.of(epochOne), leaderAndEpoch));
-        assertFalse(client.isConnected(node.idString()));
-        assertTrue(subscriptions.awaitingValidation(tp0));
-
-        // Inject an older version of the metadata response
-        final short responseVersion = 8;
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWith("dummy", 1,
-            Collections.emptyMap(), partitionCounts, tp -> null, MetadataResponse.PartitionMetadata::new, responseVersion, topicIds), false, 0L);
-        metadataFetcher.validateOffsetsIfNeeded();
-        // Offset validation is skipped
-        assertFalse(subscriptions.awaitingValidation(tp0));
-    }
-
-    @Test
-    public void testOffsetValidationResetOffsetForUndefinedEpochWithDefinedResetPolicy() {
-        testOffsetValidationWithGivenEpochOffset(
-            UNDEFINED_EPOCH, 0L, OffsetResetStrategy.EARLIEST);
-    }
-
-    @Test
-    public void testOffsetValidationResetOffsetForUndefinedOffsetWithDefinedResetPolicy() {
-        testOffsetValidationWithGivenEpochOffset(
-            2, UNDEFINED_EPOCH_OFFSET, OffsetResetStrategy.EARLIEST);
-    }
-
-    @Test
-    public void testOffsetValidationResetOffsetForUndefinedEpochWithUndefinedResetPolicy() {
-        testOffsetValidationWithGivenEpochOffset(
-            UNDEFINED_EPOCH, 0L, OffsetResetStrategy.NONE);
-    }
-
-    @Test
-    public void testOffsetValidationResetOffsetForUndefinedOffsetWithUndefinedResetPolicy() {
-        testOffsetValidationWithGivenEpochOffset(
-            2, UNDEFINED_EPOCH_OFFSET, OffsetResetStrategy.NONE);
-    }
-
-    @Test
-    public void testOffsetValidationTriggerLogTruncationForBadOffsetWithUndefinedResetPolicy() {
-        testOffsetValidationWithGivenEpochOffset(
-            1, 1L, OffsetResetStrategy.NONE);
-    }
-
-    private void testOffsetValidationWithGivenEpochOffset(int leaderEpoch,
-                                                          long endOffset,
-                                                          OffsetResetStrategy offsetResetStrategy) {
-        buildFetcher(offsetResetStrategy);
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        final int epochOne = 1;
-        final long initialOffset = 5;
-
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-            Collections.emptyMap(), partitionCounts, tp -> epochOne, topicIds), false, 0L);
-
-        // Offset validation requires OffsetForLeaderEpoch request v3 or higher
-        Node node = metadata.fetch().nodes().get(0);
-        apiVersions.update(node.idString(), NodeApiVersions.create());
-
-        Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-        subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(initialOffset, Optional.of(epochOne), leaderAndEpoch));
-
-        metadataFetcher.validateOffsetsIfNeeded();
-
-        consumerClient.poll(time.timer(Duration.ZERO));
-        assertTrue(subscriptions.awaitingValidation(tp0));
-        assertTrue(client.hasInFlightRequests());
-
-        client.respond(
-            offsetsForLeaderEpochRequestMatcher(tp0, epochOne, epochOne),
-            prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, leaderEpoch, endOffset));
-        consumerClient.poll(time.timer(Duration.ZERO));
-
-        if (offsetResetStrategy == OffsetResetStrategy.NONE) {
-            LogTruncationException thrown =
-                assertThrows(LogTruncationException.class, () -> metadataFetcher.validateOffsetsIfNeeded());
-            assertEquals(singletonMap(tp0, initialOffset), thrown.offsetOutOfRangePartitions());
-
-            if (endOffset == UNDEFINED_EPOCH_OFFSET || leaderEpoch == UNDEFINED_EPOCH) {
-                assertEquals(Collections.emptyMap(), thrown.divergentOffsets());
-            } else {
-                OffsetAndMetadata expectedDivergentOffset = new OffsetAndMetadata(
-                    endOffset, Optional.of(leaderEpoch), "");
-                assertEquals(singletonMap(tp0, expectedDivergentOffset), thrown.divergentOffsets());
-            }
-            assertTrue(subscriptions.awaitingValidation(tp0));
-        } else {
-            metadataFetcher.validateOffsetsIfNeeded();
-            assertFalse(subscriptions.awaitingValidation(tp0));
-        }
-    }
-
-    @Test
-    public void testOffsetValidationHandlesSeekWithInflightOffsetForLeaderRequest() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        final int epochOne = 1;
-
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), partitionCounts, tp -> epochOne, topicIds), false, 0L);
-
-        // Offset validation requires OffsetForLeaderEpoch request v3 or higher
-        Node node = metadata.fetch().nodes().get(0);
-        apiVersions.update(node.idString(), NodeApiVersions.create());
-
-        Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-        subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(epochOne), leaderAndEpoch));
-
-        metadataFetcher.validateOffsetsIfNeeded();
-        consumerClient.poll(time.timer(Duration.ZERO));
-        assertTrue(subscriptions.awaitingValidation(tp0));
-        assertTrue(client.hasInFlightRequests());
-
-        // While the OffsetForLeaderEpoch request is in-flight, we seek to a different offset.
-        subscriptions.seekUnvalidated(tp0, new SubscriptionState.FetchPosition(5, Optional.of(epochOne), leaderAndEpoch));
-        assertTrue(subscriptions.awaitingValidation(tp0));
-
-        client.respond(
-            offsetsForLeaderEpochRequestMatcher(tp0, epochOne, epochOne),
-            prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, 0, 0L));
-        consumerClient.poll(time.timer(Duration.ZERO));
-
-        // The response should be ignored since we were validating a different position.
-        assertTrue(subscriptions.awaitingValidation(tp0));
-    }
-
-    @Test
-    public void testOffsetValidationFencing() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        final int epochOne = 1;
-        final int epochTwo = 2;
-        final int epochThree = 3;
-
-        // Start with metadata, epoch=1
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), partitionCounts, tp -> epochOne, topicIds), false, 0L);
-
-        // Offset validation requires OffsetForLeaderEpoch request v3 or higher
-        Node node = metadata.fetch().nodes().get(0);
-        apiVersions.update(node.idString(), NodeApiVersions.create());
-
-        // Seek with a position and leader+epoch
-        Metadata.LeaderAndEpoch leaderAndEpoch = new Metadata.LeaderAndEpoch(metadata.currentLeader(tp0).leader, Optional.of(epochOne));
-        subscriptions.seekValidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(epochOne), leaderAndEpoch));
-
-        // Update metadata to epoch=2, enter validation
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), partitionCounts, tp -> epochTwo, topicIds), false, 0L);
-        metadataFetcher.validateOffsetsIfNeeded();
-        assertTrue(subscriptions.awaitingValidation(tp0));
-
-        // Update the position to epoch=3, as we would from a fetch
-        subscriptions.completeValidation(tp0);
-        SubscriptionState.FetchPosition nextPosition = new SubscriptionState.FetchPosition(
-                10,
-                Optional.of(epochTwo),
-                new Metadata.LeaderAndEpoch(leaderAndEpoch.leader, Optional.of(epochTwo)));
-        subscriptions.position(tp0, nextPosition);
-        subscriptions.maybeValidatePositionForCurrentLeader(apiVersions, tp0, new Metadata.LeaderAndEpoch(leaderAndEpoch.leader, Optional.of(epochThree)));
-
-        // Prepare offset list response from async validation with epoch=2
-        client.prepareResponse(prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, epochTwo, 10L));
-        consumerClient.pollNoWakeup();
-        assertTrue(subscriptions.awaitingValidation(tp0), "Expected validation to fail since leader epoch changed");
-
-        // Next round of validation, should succeed in validating the position
-        metadataFetcher.validateOffsetsIfNeeded();
-        client.prepareResponse(prepareOffsetsForLeaderEpochResponse(tp0, Errors.NONE, epochThree, 10L));
-        consumerClient.pollNoWakeup();
-        assertFalse(subscriptions.awaitingValidation(tp0), "Expected validation to succeed with latest epoch");
-    }
-
-    @Test
-    public void testSkipValidationForOlderApiVersion() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-
-        Map<String, Integer> partitionCounts = new HashMap<>();
-        partitionCounts.put(tp0.topic(), 4);
-
-        apiVersions.update("0", NodeApiVersions.create(ApiKeys.OFFSET_FOR_LEADER_EPOCH.id, (short) 0, (short) 2));
-
-        // Start with metadata, epoch=1
-        metadata.updateWithCurrentRequestVersion(RequestTestUtils.metadataUpdateWithIds("dummy", 1,
-                Collections.emptyMap(), partitionCounts, tp -> 1, topicIds), false, 0L);
-
-        // Request offset reset
-        subscriptions.requestOffsetReset(tp0, OffsetResetStrategy.LATEST);
-
-        // Since we have no position due to reset, no fetch is sent
-        assertEquals(0, fetcher.sendFetches());
-
-        // Still no position, ensure offset validation logic did not transition us to FETCHING state
-        assertEquals(0, fetcher.sendFetches());
-
-        // Complete reset and now we can fetch
-        metadataFetcher.resetOffsetIfNeeded(tp0, OffsetResetStrategy.LATEST,
-                new MetadataFetcher.ListOffsetData(100, 1L, Optional.empty()));
-        assertEquals(1, fetcher.sendFetches());
-    }
-
-    @Test
     public void testTruncationDetected() {
         // Create some records that include a leader epoch (1)
         MemoryRecordsBuilder builder = MemoryRecords.builder(
@@ -4700,7 +3161,8 @@ public class FetcherTest {
         builder.appendWithOffset(2L, 0L, "key".getBytes(), "value-3".getBytes());
         MemoryRecords records = builder.build();
 
-        buildFetcher();
+        buildFetcher(OffsetResetStrategy.EARLIEST, new ByteArrayDeserializer(), new ByteArrayDeserializer(),
+                Integer.MAX_VALUE, IsolationLevel.READ_UNCOMMITTED);
         assignFromUser(singleton(tp0));
 
         // Initialize the epoch=2
@@ -4719,6 +3181,15 @@ public class FetcherTest {
         subscriptions.seekValidated(tp0, new SubscriptionState.FetchPosition(0, Optional.of(1), leaderAndEpoch));
 
         // Check for truncation, this should cause tp0 to go into validation
+        MetadataFetcher metadataFetcher = new MetadataFetcher(new LogContext(),
+                consumerClient,
+                metadata,
+                subscriptions,
+                time,
+                retryBackoffMs,
+                requestTimeoutMs,
+                IsolationLevel.READ_UNCOMMITTED,
+                apiVersions);
         metadataFetcher.validateOffsetsIfNeeded();
 
         // No fetches sent since we entered validation
@@ -4983,89 +3454,6 @@ public class FetcherTest {
         assertThrows(KafkaException.class, this::fetchedRecords);
     }
 
-    @Test
-    public void testBeginningOffsets() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.EARLIEST_TIMESTAMP, 2L));
-        assertEquals(singletonMap(tp0, 2L), metadataFetcher.beginningOffsets(singleton(tp0), time.timer(5000L)));
-    }
-
-    @Test
-    public void testBeginningOffsetsDuplicateTopicPartition() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.EARLIEST_TIMESTAMP, 2L));
-        assertEquals(singletonMap(tp0, 2L), metadataFetcher.beginningOffsets(asList(tp0, tp0), time.timer(5000L)));
-    }
-
-    @Test
-    public void testBeginningOffsetsMultipleTopicPartitions() {
-        buildFetcher();
-        Map<TopicPartition, Long> expectedOffsets = new HashMap<>();
-        expectedOffsets.put(tp0, 2L);
-        expectedOffsets.put(tp1, 4L);
-        expectedOffsets.put(tp2, 6L);
-        assignFromUser(expectedOffsets.keySet());
-        client.prepareResponse(listOffsetResponse(expectedOffsets, Errors.NONE, ListOffsetsRequest.EARLIEST_TIMESTAMP, ListOffsetsResponse.UNKNOWN_EPOCH));
-        assertEquals(expectedOffsets, metadataFetcher.beginningOffsets(asList(tp0, tp1, tp2), time.timer(5000L)));
-    }
-
-    @Test
-    public void testBeginningOffsetsEmpty() {
-        buildFetcher();
-        assertEquals(emptyMap(), metadataFetcher.beginningOffsets(emptyList(), time.timer(5000L)));
-    }
-
-    @Test
-    public void testEndOffsets() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.LATEST_TIMESTAMP, 5L));
-        assertEquals(singletonMap(tp0, 5L), metadataFetcher.endOffsets(singleton(tp0), time.timer(5000L)));
-    }
-
-    @Test
-    public void testEndOffsetsDuplicateTopicPartition() {
-        buildFetcher();
-        assignFromUser(singleton(tp0));
-        client.prepareResponse(listOffsetResponse(tp0, Errors.NONE, ListOffsetsRequest.LATEST_TIMESTAMP, 5L));
-        assertEquals(singletonMap(tp0, 5L), metadataFetcher.endOffsets(asList(tp0, tp0), time.timer(5000L)));
-    }
-
-    @Test
-    public void testEndOffsetsMultipleTopicPartitions() {
-        buildFetcher();
-        Map<TopicPartition, Long> expectedOffsets = new HashMap<>();
-        expectedOffsets.put(tp0, 5L);
-        expectedOffsets.put(tp1, 7L);
-        expectedOffsets.put(tp2, 9L);
-        assignFromUser(expectedOffsets.keySet());
-        client.prepareResponse(listOffsetResponse(expectedOffsets, Errors.NONE, ListOffsetsRequest.LATEST_TIMESTAMP, ListOffsetsResponse.UNKNOWN_EPOCH));
-        assertEquals(expectedOffsets, metadataFetcher.endOffsets(asList(tp0, tp1, tp2), time.timer(5000L)));
-    }
-
-    @Test
-    public void testEndOffsetsEmpty() {
-        buildFetcher();
-        assertEquals(emptyMap(), metadataFetcher.endOffsets(emptyList(), time.timer(5000L)));
-    }
-
-    private MockClient.RequestMatcher offsetsForLeaderEpochRequestMatcher(
-        TopicPartition topicPartition,
-        int currentLeaderEpoch,
-        int leaderEpoch
-    ) {
-        return request -> {
-            OffsetsForLeaderEpochRequest epochRequest = (OffsetsForLeaderEpochRequest) request;
-            OffsetForLeaderPartition partition = offsetForLeaderPartitionMap(epochRequest.data())
-                .get(topicPartition);
-            return partition != null
-                && partition.currentLeaderEpoch() == currentLeaderEpoch
-                && partition.leaderEpoch() == leaderEpoch;
-        };
-    }
-
     private OffsetsForLeaderEpochResponse prepareOffsetsForLeaderEpochResponse(
         TopicPartition topicPartition,
         Errors error,
@@ -5081,69 +3469,6 @@ public class FetcherTest {
                 .setLeaderEpoch(leaderEpoch)
                 .setEndOffset(endOffset))));
         return new OffsetsForLeaderEpochResponse(data);
-    }
-
-    private Map<TopicPartition, OffsetForLeaderPartition> offsetForLeaderPartitionMap(
-        OffsetForLeaderEpochRequestData data
-    ) {
-        Map<TopicPartition, OffsetForLeaderPartition> result = new HashMap<>();
-        data.topics().forEach(topic ->
-            topic.partitions().forEach(partition ->
-                result.put(new TopicPartition(topic.topic(), partition.partition()), partition)));
-        return result;
-    }
-
-    private MockClient.RequestMatcher listOffsetRequestMatcher(final long timestamp) {
-        return listOffsetRequestMatcher(timestamp, ListOffsetsResponse.UNKNOWN_EPOCH);
-    }
-
-    private MockClient.RequestMatcher listOffsetRequestMatcher(final long timestamp, final int leaderEpoch) {
-        // matches any list offset request with the provided timestamp
-        return body -> {
-            ListOffsetsRequest req = (ListOffsetsRequest) body;
-            ListOffsetsTopic topic = req.topics().get(0);
-            ListOffsetsPartition partition = topic.partitions().get(0);
-            return tp0.topic().equals(topic.name())
-                    && tp0.partition() == partition.partitionIndex()
-                    && timestamp == partition.timestamp()
-                    && leaderEpoch == partition.currentLeaderEpoch();
-        };
-    }
-
-    private ListOffsetsResponse listOffsetResponse(Errors error, long timestamp, long offset) {
-        return listOffsetResponse(tp0, error, timestamp, offset);
-    }
-
-    private ListOffsetsResponse listOffsetResponse(TopicPartition tp, Errors error, long timestamp, long offset) {
-        return listOffsetResponse(tp, error, timestamp, offset, ListOffsetsResponse.UNKNOWN_EPOCH);
-    }
-
-    private ListOffsetsResponse listOffsetResponse(TopicPartition tp, Errors error, long timestamp, long offset, int leaderEpoch) {
-        Map<TopicPartition, Long> offsets = new HashMap<>();
-        offsets.put(tp, offset);
-        return listOffsetResponse(offsets, error, timestamp, leaderEpoch);
-    }
-
-    private ListOffsetsResponse listOffsetResponse(Map<TopicPartition, Long> offsets, Errors error, long timestamp, int leaderEpoch) {
-        Map<String, List<ListOffsetsPartitionResponse>> responses = new HashMap<>();
-        for (Map.Entry<TopicPartition, Long> entry : offsets.entrySet()) {
-            TopicPartition tp = entry.getKey();
-            responses.putIfAbsent(tp.topic(), new ArrayList<>());
-            responses.get(tp.topic()).add(new ListOffsetsPartitionResponse()
-                                .setPartitionIndex(tp.partition())
-                                .setErrorCode(error.code())
-                                .setOffset(entry.getValue())
-                                .setTimestamp(timestamp)
-                                .setLeaderEpoch(leaderEpoch));
-        }
-        List<ListOffsetsTopicResponse> topics = new ArrayList<>();
-        for (Map.Entry<String, List<ListOffsetsPartitionResponse>> response : responses.entrySet()) {
-            topics.add(new ListOffsetsTopicResponse()
-                    .setName(response.getKey())
-                    .setPartitions(response.getValue()));
-        }
-        ListOffsetsResponseData data = new ListOffsetsResponseData().setTopics(topics);
-        return new ListOffsetsResponse(data);
     }
 
     private FetchResponse fetchResponseWithTopLevelError(TopicIdPartition tp, Errors error, int throttleTime) {
@@ -5226,25 +3551,6 @@ public class FetcherTest {
         return FetchResponse.of(Errors.NONE, throttleTime, INVALID_SESSION_ID, new LinkedHashMap<>(partitions));
     }
 
-    private MetadataResponse newMetadataResponse(String topic, Errors error) {
-        List<MetadataResponse.PartitionMetadata> partitionsMetadata = new ArrayList<>();
-        if (error == Errors.NONE) {
-            Optional<MetadataResponse.TopicMetadata> foundMetadata = initialUpdateResponse.topicMetadata()
-                    .stream()
-                    .filter(topicMetadata -> topicMetadata.topic().equals(topic))
-                    .findFirst();
-            foundMetadata.ifPresent(topicMetadata -> {
-                partitionsMetadata.addAll(topicMetadata.partitionMetadata());
-            });
-        }
-
-        MetadataResponse.TopicMetadata topicMetadata = new MetadataResponse.TopicMetadata(error, topic, false,
-                partitionsMetadata);
-        List<Node> brokers = new ArrayList<>(initialUpdateResponse.brokers());
-        return RequestTestUtils.metadataResponse(brokers, initialUpdateResponse.clusterId(),
-                initialUpdateResponse.controller().id(), Collections.singletonList(topicMetadata));
-    }
-
     /**
      * Assert that the {@link Fetcher#collectFetch() latest fetch} does not contain any
      * {@link Fetch#records() user-visible records}, did not
@@ -5282,12 +3588,6 @@ public class FetcherTest {
                               Deserializer<?> valueDeserializer) {
         buildFetcher(OffsetResetStrategy.EARLIEST, keyDeserializer, valueDeserializer,
                 Integer.MAX_VALUE, IsolationLevel.READ_UNCOMMITTED);
-    }
-
-    private void buildFetcher(OffsetResetStrategy offsetResetStrategy) {
-        buildFetcher(new MetricConfig(), offsetResetStrategy,
-            new ByteArrayDeserializer(), new ByteArrayDeserializer(),
-            Integer.MAX_VALUE, IsolationLevel.READ_UNCOMMITTED);
     }
 
     private <K, V> void buildFetcher(OffsetResetStrategy offsetResetStrategy,
@@ -5348,7 +3648,7 @@ public class FetcherTest {
                 metricsRegistry,
                 time,
                 isolationLevel,
-                apiVersions);
+                apiVersions));
         metadataFetcher = new MetadataFetcher(logContext,
                 consumerClient,
                 metadata,
@@ -5357,7 +3657,7 @@ public class FetcherTest {
                 retryBackoffMs,
                 requestTimeoutMs,
                 isolationLevel,
-                apiVersions));
+                apiVersions);
     }
 
     private void buildDependencies(MetricConfig metricConfig,
